requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
customtkinter==5.2.2
darkdetect==0.8.0
idna==3.10
packaging==24.2
pillow==11.1.0
pycryptodome==3.22.0
requests==2.32.3
urllib3==2.3.0